requirements.txt
pydantic-ai-slim[openai,mcp]
fastmcp
sqlalchemy 
psycopg2-binary 
faker
